Move logic to a lock module
