+++ services/db.js
@@ -85,9 +85,7 @@ module.exports.put = put;
 module.exports.get = get;
 module.exports.del = del;
 module.exports.batch = batch;
-
 module.exports.applyLock = redis.applyLock;
-
 module.exports.raw = postgres.raw;
 module.exports.putMeta = postgres.putMeta;
 module.exports.getMeta = postgres.getMeta;
